Account with default Controller Configuration

+++ account-service/src/main/java/com/example/demo/rest/dto/AccountDTO.java
@@ -23,7 +23,7 @@ public class AccountDTO {
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
     private Currency currency;
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
-    private Long balance;
+    private Double balance;
 
 
     public String getAccountNumber() {
